AI Agent Project/requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
google-ai-generativelanguage==0.7.0
python-dotenv==1.0.1
duckduckgo-search==5.2.2
wikipedia==1.4.0
